Extract parsing code from `SettingsImporter`
